+++ Project1/magic.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "fstream"
 
-uint64_t get_magic_offset(std::fstream& fs, uint8_t* magic, int magic_len);
+size_t get_magic_offset(std::fstream& fs, uint8_t* magic, int magic_len);
